Fix error with fucntion name when uploading an image file

+++ src/providers/aws/client/codezip.py
@@ -40,15 +40,14 @@ def add_mandatory_files(function_name, env_vars):
     env_vars['UDOCKER_BIN'] = "/var/task/udocker/bin/"
 
 def create_code_zip(function_name, env_vars, script=None, extra_payload=None, image_id=None, image_file=None, deployment_bucket=None, file_key=None):
-    
-    # Delete created temporal files
-    if os.path.isdir(scar_temporal_folder):
-        shutil.rmtree(scar_temporal_folder)    
-    
+    clean_tmp_folders()
     add_mandatory_files(function_name, env_vars)
     create_udocker_files()
     if (image_id and image_id != "") and (deployment_bucket and deployment_bucket != ""):
         download_udocker_image(image_id, env_vars)
+    
+    if image_file and image_file != "":
+        prepare_udocker_image(image_file, env_vars)
         
     if script:
         shutil.copy(script, scar_temporal_folder + "init_script.sh")
@@ -63,8 +62,13 @@ def create_code_zip(function_name, env_vars, script=None, extra_payload=None, im
         
     if deployment_bucket and deployment_bucket != "":
         upload_file_to_S3_bucket(zip_file_path, deployment_bucket, file_key)
-    # Delete created temporal files        
-    shutil.rmtree(scar_temporal_folder)
+    
+    #clean_tmp_folders()
+    
+def clean_tmp_folders():
+    # Delete created temporal files
+    if os.path.isdir(scar_temporal_folder):
+        shutil.rmtree(scar_temporal_folder)
     
 def zip_scar_folder():
     execute_command(["zip", "-r9y", zip_file_path, "."], cmd_wd=scar_temporal_folder, cli_msg="Creating function package")
@@ -88,19 +92,33 @@ def restore_udocker_env():
 def execute_command(command, cmd_wd=None, cli_msg=None):
     cmd_out = subprocess.check_output(command, cwd=cmd_wd).decode("utf-8")
     logger.info(cli_msg, cmd_out)
+    return cmd_out[:-1]
     
 def create_udocker_files():
     set_tmp_udocker_env()
     execute_command(["python3", udocker_exec, "help"], cli_msg="Setting udocker environment")
     restore_udocker_env()
+
+def prepare_udocker_image(image_file, env_vars):
+    set_tmp_udocker_env()
+    shutil.copy(image_file, "/tmp/udocker_image.tar.gz")
+    cmd_out = execute_command(["python3", udocker_exec, "load", "-i", "/tmp/udocker_image.tar.gz"], cli_msg="Loading image file")
+    create_udocker_container(cmd_out)
+    env_vars['IMAGE_ID'] = cmd_out
+    env_vars['UDOCKER_REPOS'] = "/var/task/udocker/repos/"
+    env_vars['UDOCKER_LAYERS'] = "/var/task/udocker/layers/"    
+    restore_udocker_env()
+
+def create_udocker_container(image_id):
+    if(utils.get_tree_size(scar_temporal_folder) < MAX_S3_PAYLOAD_SIZE/2):
+        execute_command(["python3", udocker_exec, "create", "--name=lambda_cont", image_id], cli_msg="Creating container structure")
+    if(utils.get_tree_size(scar_temporal_folder) > MAX_S3_PAYLOAD_SIZE):
+        shutil.rmtree(scar_temporal_folder + "udocker/containers/")    
     
 def download_udocker_image(image_id, env_vars):
     set_tmp_udocker_env()
     execute_command(["python3", udocker_exec, "pull", image_id], cli_msg="Downloading container image")
-    if(utils.get_tree_size(scar_temporal_folder) < MAX_S3_PAYLOAD_SIZE/2):
-        execute_command(["python3", udocker_exec, "create", "--name=lambda_cont", image_id], cli_msg="Creating container structure")
-    if(utils.get_tree_size(scar_temporal_folder) > MAX_S3_PAYLOAD_SIZE):
-        shutil.rmtree(scar_temporal_folder + "udocker/containers/")
+    create_udocker_container(image_id)
     env_vars['UDOCKER_REPOS'] = "/var/task/udocker/repos/"
     env_vars['UDOCKER_LAYERS'] = "/var/task/udocker/layers/"
     restore_udocker_env()
